Fixed behavior with follow-giveaways
Bot now checks if a follow is required to enter the giveaway. If so, it skips the giveaway and continues with the next one. Does not make an entry in the database for these, doesn't seem necessary since it's going to close them anyway. Code could be added somewhere to allow entry in these follow giveaways if one is so inclined.